refactor: 댓글섹션 컴포넌트 분리 및 콘텐츠 파싱 유틸 함수 분리
